Thursday Changes - Reversal Fixed, Doubles instead of Integers
-- nan == nan switched to isnan() to allow the robot to reverse on vertical lines because their slope is nan
-- all ints in PIDMover and PIDCalc switched to doubles to increase accuracy

+++ include/main.h
@@ -102,7 +102,7 @@ void PIDMover(Coordinate goalPosition, bool reverse = false,                 std
 void PIDTurner(int setPoint, int direction,                 std::vector<std::function<void(void)>> custom = {}, std::vector<int> executeAt = {});
 void PIDArc(int chordLength, int maxDist, int direction,                std::vector<std::function<void(void)>> custom = {}, std::vector<int> executeAt = {});
 
-PIDReturn PIDCalc(int distanceMoved, int setPoint, bool isPositive, ConstantContainer constants, PIDReturn lastCycle);
+PIDReturn PIDCalc(double distanceMoved, double setPoint, bool isPositive, ConstantContainer constants, PIDReturn lastCycle);
 
 // autons.cpp
 void blueGoalside(void);
